05-projects/eval-automation/requirements.txt
python-dotenv==1.0.1
langfuse==2.3.1
requests==2.32.3